story(issue-34): add Kafka.RedpandaCluster + split kafka module into focused files
Adds Kafka.RedpandaCluster() backed by redpandadata/redpanda — single-node,
PLAINTEXT + TLS, dedicated *RedpandaCluster and *RedpandaServerSecurity
types because the config layer (rpk redpanda start, redpanda.yaml, PEM
certs) doesn't share the KAFKA_* env-var contract the Apache/Confluent
distros use. Wire protocol matches Kafka, so RedpandaCluster.Client() reuses
the existing *Client unchanged. Verified with PLAINTEXT and TLS round-trip
tests; full suite green.

Also splits the now-2000+-line main.go in both the kafka module and its
tests module into per-responsibility files (security.go, cluster_kafka.go,
internal_ca.go, cluster_redpanda.go, client.go, util.go; tests split by
distro into tests_native/_apache/_confluent/_redpanda + helpers.go), with
the file map documented in each module's package comment.
